refactor(node-pkg): SSH consolidation #923
