+++ k2s/cmd/k2s/cmd/node/exec/exec.go
@@ -47,7 +47,7 @@ func NewCmd() *cobra.Command {
 	cmd.MarkFlagRequired(commandFlag)
 
 	cmd.Flags().Uint16P(portFlag, "p", ssh.DefaultPort, "Port for remote connection")
-	cmd.Flags().String(timeoutFlag, "30s", "Connection timeout, e.g. '1m20s', allowed time units are 'ns', 'us' (or 'µs'), 'ms', 's', 'm', 'h'")
+	cmd.Flags().String(timeoutFlag, ssh.DefaultTimeout.String(), "Connection timeout, e.g. '1m20s', allowed time units are 'ns', 'us' (or 'µs'), 'ms', 's', 'm', 'h'")
 
 	cmd.Flags().SortFlags = false
 	cmd.Flags().PrintDefaults()
