+++ k2s/internal/core/node/ssh/ssh.go
@@ -4,9 +4,11 @@
 package ssh
 
 import (
+	"errors"
 	"fmt"
 	"log/slog"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"time"
 
@@ -22,8 +24,9 @@ type ConnectionOptions struct {
 }
 
 const (
-	SshPubKeyName        = sshKeyName + ".pub"
-	DefaultPort   uint16 = 22
+	SshPubKeyName         = sshKeyName + ".pub"
+	DefaultPort    uint16 = 22
+	DefaultTimeout        = 30 * time.Second
 
 	sshKeyName    = "id_rsa"
 	sshSubDirName = "k2s"
@@ -63,4 +66,36 @@ func Connect(options ConnectionOptions) (*ssh.Client, error) {
 
 	slog.Debug("Connected via SSH", "ip", options.IpAddress)
 	return sshClient, nil
+}
+
+func ConnectInteractively(options ConnectionOptions) error {
+	timeoutOption := fmt.Sprintf("ConnectTimeout=%d", int(options.Timeout.Seconds()))
+	port := fmt.Sprintf("%d", options.Port)
+	remote := fmt.Sprintf("%s@%s", options.RemoteUser, options.IpAddress)
+
+	cmd := exec.Command("ssh.exe", "-tt", "-o", "StrictHostKeyChecking=no", "-o", timeoutOption, "-i", options.SshKeyPath, "-p", port, remote)
+
+	slog.Debug("Executing ssh.exe", "command", cmd.String())
+
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Start(); err != nil {
+		return fmt.Errorf("failed to start ssh.exe: %w", err)
+	}
+
+	if err := cmd.Wait(); err != nil {
+		var exitErr *exec.ExitError
+		if errors.As(err, &exitErr) {
+			exitCode := exitErr.ExitCode()
+			if exitCode == 255 {
+				return fmt.Errorf("failed to execute ssh.exe: %w", err)
+			}
+			slog.Debug("failed to execute ssh.exe", "exit-code", exitErr.ExitCode())
+			return nil
+		}
+		return fmt.Errorf("failed to wait for ssh.exe execution: %w", err)
+	}
+	return nil
 }
