+++ k2s/internal/core/users/users.go
@@ -10,7 +10,9 @@ import (
 
 	"github.com/samber/lo"
 	"github.com/siemens-healthineers/k2s/internal/core/config"
-	kssh "github.com/siemens-healthineers/k2s/internal/core/node/ssh"
+	"github.com/siemens-healthineers/k2s/internal/core/node"
+	"github.com/siemens-healthineers/k2s/internal/core/node/copy"
+	"github.com/siemens-healthineers/k2s/internal/core/node/ssh"
 	"github.com/siemens-healthineers/k2s/internal/core/users/acl"
 	"github.com/siemens-healthineers/k2s/internal/core/users/common"
 	"github.com/siemens-healthineers/k2s/internal/core/users/fs"
@@ -20,8 +22,6 @@ import (
 	"github.com/siemens-healthineers/k2s/internal/core/users/k8s/kubeconfig"
 	"github.com/siemens-healthineers/k2s/internal/core/users/nodes"
 	"github.com/siemens-healthineers/k2s/internal/core/users/nodes/keygen"
-	"github.com/siemens-healthineers/k2s/internal/core/users/nodes/scp"
-	"github.com/siemens-healthineers/k2s/internal/core/users/nodes/ssh"
 	"github.com/siemens-healthineers/k2s/internal/core/users/winusers"
 )
 
@@ -62,18 +62,22 @@ func NewUsersManagement(cfg *config.Config, cmdExecutor common.CmdExecutor, user
 		exec: cmdExecutor,
 	}
 
-	sshKeyPath := kssh.SshKeyPath(cfg.Host.SshDir)
-	remoteUser := nodes.DetermineSshRemoteUser(controlePlaneCfg.IpAddress)
-	sshExec := ssh.NewSsh(cmdExecutor, sshKeyPath, remoteUser)
-	scpExec := scp.NewScp(cmdExecutor, sshKeyPath, remoteUser)
+	sshOptions := ssh.ConnectionOptions{
+		RemoteUser: "remote",
+		IpAddress:  controlePlaneCfg.IpAddress,
+		Port:       ssh.DefaultPort,
+		SshKeyPath: ssh.SshKeyPath(cfg.Host.SshDir),
+		Timeout:    ssh.DefaultTimeout,
+	}
+
 	fileSystem := fs.NewFileSystem()
 	keygenExec := keygen.NewSshKeyGen(cmdExecutor, fileSystem)
 	aclExec := acl.NewAcl(cmdExecutor)
 	restClient := http.NewRestClient()
 	kubeconfigReader := kubeconfig.NewKubeconfigReader()
-	controlPlaneAccess := nodes.NewControlPlaneAccess(fileSystem, keygenExec, sshExec, scpExec, aclExec, cfg.Host.SshDir, controlePlaneCfg.IpAddress)
+	controlPlaneAccess := nodes.NewControlPlaneAccess(fileSystem, keygenExec, node.Exec, copy.Copy, sshOptions, aclExec, cfg.Host.SshDir, controlePlaneCfg.IpAddress)
 	clusterAccess := cluster.NewClusterAccess(restClient)
-	k8sAccess := k8s.NewK8sAccess(sshExec, scpExec, fileSystem, clusterAccess, kubeconfigWriterFactory, kubeconfigReader, cfg.Host.KubeConfigDir)
+	k8sAccess := k8s.NewK8sAccess(node.Exec, copy.Copy, sshOptions, fileSystem, clusterAccess, kubeconfigWriterFactory, kubeconfigReader, cfg.Host.KubeConfigDir)
 	userAdder := NewWinUserAdder(controlPlaneAccess, k8sAccess, CreateK2sUserName)
 
 	return &usersManagement{
